splitted calendar into components
