+++ src/components/ChoosedMonth/CalendarTable/testCalendar.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { useCalendar } from '../useCalendar';
+import { CalendarTable } from '../CalendarTable';
+import { PeriodPaginator } from '../../CalendarToolbar/PeriodPaginator';
+
+export const CalendarTableTest = () => {
+  const { calendarProps, title, prevButtonProps, nextButtonProps, state } =
+    useCalendar();
+
+  return (
+    <div {...calendarProps} className=" w-3/4 m-auto mt-52">
+      <PeriodPaginator
+        title={title}
+        prevButtonProps={prevButtonProps}
+        nextButtonProps={nextButtonProps}
+      />
+
+      <CalendarTable state={state} />
+    </div>
+  );
+};
