+++ src/components/ChoosedMonth/index.js
@@ -1,1 +1,2 @@
 export * from './ChoosedMonth';
+export * from './useCalendar';
