+++ src/pages/CalendarPage/CalendarPage.jsx
@@ -1,15 +1,28 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { CalendarToolbar, ChoosedDay, ChoosedMonth } from '../../components';
+import { useCalendar } from '../../components/ChoosedMonth';
 
 export const CalendarPage = () => {
+  const {
+    calendarProps,
+    title,
+    prevButtonProps,
+    nextButtonProps,
+    state,
+    daysOfWeekLabels,
+    dateFormatter,
+  } = useCalendar();
+
   // consume hook
   const isMonthView = true;
+  const properties = { state, calendarProps, daysOfWeekLabels, dateFormatter };
 
   return (
     <div>
-      <CalendarToolbar />
+      <CalendarToolbar {...{ title, prevButtonProps, nextButtonProps }} />
 
-      {isMonthView ? <ChoosedMonth /> : <ChoosedDay />}
+      {isMonthView && <ChoosedMonth {...properties} />}
+      {!isMonthView && <ChoosedDay {...properties} />}
     </div>
   );
 };
